perf: remove WordsContext
